refactor(core): split engine unit tests by concern
Move engine/tests.rs into tests/{load,stream,transcribe,backends,mock}.rs
so load/stream/transcribe coverage sits next to the impl files they cover.
